AuthController@loginのテスト追加

+++ app/Services/AuthService.php
@@ -61,7 +61,7 @@ class AuthService extends Service
             return [
                 'access_token' => $jwt,
                 'token_type'   => 'bearer',
-                'expires_in'   => env('JWT_TTL')
+                'expires_in'   => (int) env('JWT_TTL'),
             ];
 
         } catch ( Exception $e ) {
